requirements.txt
azure-iot-hub
aiohttp